scpstApp/requirements.txt
flask
selenium
webdriver-manager
google-cloud-storage
paddleocr
paddlepaddle
opencv-python
flask-cors
flask-login
firebase-admin